pyproject.toml
[project]
name = "bandcamp-newsfeed-rss"
version = "0"
description = "FastAPI RSS Feed Generator for your Bandcamp newsfeed "
readme = "README.md"
requires-python = "==3.14.*"
dependencies = [
    "beautifulsoup4",
    "curl-cffi",
    "fastapi",
    "feedgen",
    "pydantic",
    "python-dotenv",
    "tzdata",
    "uvicorn",
    "bandcamp-async-api>=0.0.9",  # 0.0.9 introduces get_feed
]

[dependency-groups]
dev = [
    "pre-commit-uv",
    "ruff",
    "pytest-asyncio",
    "httpx",
    "respx",
    "pytest-cov",
]


[tool.pytest.ini_options]
minversion = "8.3.5"
addopts = "-ra -q -m 'not e2e'"
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
markers = [
    "e2e: marks tests as end-to-end tests that require real network access (deselect with '-m \"not e2e\"')",
]
